fix install scripts

+++ zsh-install.sh
@@ -5,11 +5,6 @@ command_exists() {
   command -v "$1" >/dev/null 2>&1
 }
 
-# Function to check if a package is installed
-package_installed() {
-  dpkg -s "$1" >/dev/null 2>&1
-}
-
 # Function to handle errors
 handle_error() {
   echo "Error: $1"
@@ -23,25 +18,12 @@ sudo apt-get upgrade -y || handle_error "Failed to upgrade packages."
 # Install prerequisites
 prerequisites=("curl" "git" "zsh")
 for package in "${prerequisites[@]}"; do
-    sudo apt-get install "$package" -y
+    if ! command_exists "$package"; then
+      sudo apt-get install "$package" -y
+    else
+      echo "$package ja está instalado."
 done
 
-# Set zsh theme
-sed -i 's/ZSH_THEME=.*/ZSH_THEME="jonathan"/' ~/.zshrc
-
-# Install Zinit
-bash -c "$(curl --fail --show-error --silent --location https://raw.githubusercontent.com/zdharma-continuum/zinit/HEAD/scripts/install.sh)" || handle_error "Failed to install Zinit."
-
-# Configure zinit
-echo "
-# Zinit start
-zinit light zdharma-continuum/fast-syntax-highlighting
-zinit light zsh-users/zsh-autosuggestions
-zinit light zsh-users/zsh-completions
-# Zinit end
-" >> ~/.zshrc
-
-echo "Zinit setup completed."
 
 # Install Oh My Zsh
 OH_MY_ZSH_DIR="~/.oh-my-zsh"
@@ -50,5 +32,3 @@ if [ ! -d "$OH_MY_ZSH_DIR" ]; then
 else
   echo "Oh My Zsh is already installed."
 fi
-
-echo "ZSH setup completed."
